monitor_service/requirements.txt
fastapi
uvicorn
pydantic
motor
pymongo
python-dotenv